added login function and Authorize middleware

+++ App/Controllers/ListingController.php
@@ -3,6 +3,7 @@
 namespace App\Controllers;
 
 use Framework\Database;
+use Framework\Session;
 use Framework\Validation;
 
 class ListingController
@@ -16,7 +17,7 @@ class ListingController
 
     public function index()
     {
-        $listings = $this->db->query("SELECT * FROM listings")->fetchAll();
+        $listings = $this->db->query("SELECT * FROM listings ORDER BY created_at DESC")->fetchAll();
         loadView('listings/index', [
             'listings' => $listings
         ]);
@@ -65,7 +66,7 @@ class ListingController
         $aloewdField = ['title','description','salary', 'tag','company','address','city','state','phone','email','requriements','benefits'];
         $newListingData = array_intersect_key($_POST,array_flip($aloewdField));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
         $newListingData = array_map('sanitize',$newListingData);
 
         $requriedFields = ['title','description','salary','city','state','email'];
@@ -124,6 +125,11 @@ class ListingController
         if (!$listing) {
             ErrorController::notFound('listing not found');
             return;
+        }
+        // Authorization
+        if (Session::get('user')['id']!==$listing->user_id) {
+            $_SESSION['error_message'] = 'You are not authorized to delete this listing';
+            return redirect('/listings/'.$listing->id);
         }
 
         $this->db->query("DELETE FROM listings WHERE id = :id",$params);
